requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.3.1
defusedxml==0.7.1
dj-rest-auth==1.1.0
Django==3.2
django-allauth==0.42.0
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
